Add CMake support for bnk project with glad library
Created a new CMakeLists file for the overall `bnk` project and also included the `glad` library necessary for graphics rendering. This change is required to build the application and link appropriate dependencies, ensuring we are using consistent library versions across different machines. The glfw package has been specified as well for windowing setup and interactions.