Add extended call function method to Vm
Added a new method, `call_function_ex`, to the Virtual Machine (Vm) in both the header (vm.h) and source (vm.cpp) files. This method extends the existing `call_function` by executing the function and returning the result. This was necessary to allow for more complex functionality and operations outside the VM.