Add INSTANCE macro and extend trait system in the code
The INSTANCE macro was added to all classes defined in object.h.
This macro allows easier type checking and debugging.
Furthermore, the trait system was extended in traits.hpp. With these changes,
it is possible to work with standard types and to automatically convert them
to internal representation and vice versa.
Functionality was added to unwrap function arguments and to handle native functions and methods as well.
This is important for connecting with external libraries and for creating bindings in the future.