Add error check for non-result return in error functions
A section of code has been modified to check if a
return statement in an error function is of Result type,
if not it would throw a runtime error.
This change improves the safety and correctness of the language
by enacting type checking policies.

Additional changes include the addition of is_digit() and is_alpha()
methods to the String type along with some tweaks to the build and
package scripts for better library handling.