Update Http, Engine and Compiler code
In this commit, the Http, Engine and Compiler sections were updated.

In Http, the Request class was updated to take 'httplib::Request'
by value to avoid lifetime issues. The Response class was also updated to indicate
a response body and status getters in http_client.

For Engine, significant upgrades were made to the REPL which included adding history and command support,
tab completion, syntax highlighting and collision detection on imported modules.
However, there are some known bugs such as failing compilations
and crashes that can cause globals and scopes to mismatch and the
history section does not persist across REPL sessions.

Lastly, in the Compiler, changes were made on parser.cpp, lexer.cpp and bytecode.cpp
such as adding new operations and the introduction of a flag to indicate that disassembly is enabled.
It also added a function to remove unnecessary objects from the global scope to fix the aforementioned collisions in the REPL.