Add support for unpacking variables in for loops
Implement the ability to unpack variables in for loops.
This change modifies the `CodeGen`, `Parser`, `Resolver`,
`vm`, and `code` classes for supporting the unpacking
construct in for loops, enhancing flexibility for developers.
Variable declarations in the for loop are now contained in a vector,
providing support for listing multiple variables.