Refactored code for better cross-platform compatibility
This commit includes refactoring of code to work better on Linux systems.
It includes changes such as adding condition to check for WIN32 for windows specific includes,
changing getter and setter function names to getter_fn and setter_fn.
In build.cpp, added conditions to alternate between .dll and .so files depending on the platform.