Optimize attribute retrieval in codegen and vm
The attribute retrieval in Codegen has been reverted to use
'make_string_cache' instead of 'add_or_get_identifier' to. Also,
the error handling for 'get_attr' in VM has been modified.
Instead of throwing a runtime error on failing to get an attribute
of a NativeInstance, it falls back to '__get_attr__'
if available, preventing unnecessary program termination.